refactor: app과 카드정보 폼에서 훅 분리하기
