+++ src/cardData.ts
@@ -11,6 +11,15 @@ export const cardCoList: CardCo[] = [
   "kakao",
 ];
 
+export const initialCardInputComplete = {
+  cardCo: false,
+  cardNumber: false,
+  expirationDate: false,
+  owner: false,
+  securityCode: false,
+  password: false,
+};
+
 export const initialCard: CreditCard = {
   nickName: "",
   owner: "",
